Updated tooltip examples for react 16
THis should close #563
