+++ docs/src/components/Components/Tooltips/CustomExamples/code.js
@@ -1,0 +1,17 @@
+import CustomExamples from '!!raw-loader!./index.jsx';
+import TooltipFontIcon from '!!raw-loader!./TooltipFontIcon.jsx';
+import TooltipLink from '!!raw-loader!./TooltipLink.jsx';
+
+export default `/* CustomExamples.jsx */
+${CustomExamples}
+\`\`\`
+
+\`\`\`jsx
+/* TooltipFontIcon.jsx */
+${TooltipFontIcon}
+\`\`\`
+
+\`\`\`jsx
+/* TooltipLink.jsx */
+${TooltipLink}
+`;
